fix cache_dir handling
in previous implementation, dukkha.CacheDir was not set until all renderers
resolved, renderer cache dirs are always relative to the dukkha.CacheDir.
when renderer definition is dynamic, dukkha will use system cwd as cache_dir
that is not the expected behavior.

In this fix, we replaced the config option global.cache_dir with env
DUKKHA_CACHE_DIR.

Close #162